trip-chat-service/pyproject.toml
[project]
name = "chat"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk>=0.3.13",
    "grpcio>=1.75.1",
    "litestar[opentelemetry,standard]>=2.18.0",
    "nacos-sdk-python>=2.0.9",
    "opentelemetry-distro[otlp]>=0.58b0",
    "pydantic>=2.12.0",
    "pydantic-ai-slim[openai]>=1.19.0",
    "pydantic-settings>=2.11.0",
    "pymongo>=4.15.3",
    "tripsphere",
]

[project.scripts]
boot = "chat.cli.main:run_cli"

[build-system]
requires = ["uv_build>=0.9.1,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "grpcio-tools>=1.75.1",
    "mypy>=1.18.2",
    "mypy-protobuf",
    "pyright>=1.1.406",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-mock>=3.15.1",
    "types-grpcio>=1.0.0.20251009",
    "v2-stubs>=2.0.9",
]

[tool.uv.sources]
tripsphere = { workspace = true }
v2-stubs = { path = "../libs/v2-stubs", editable = true }
mypy-protobuf = { git = "https://github.com/nipunn1313/mypy-protobuf", branch = "main" }

[tool.uv.workspace]
members = ["libs/tripsphere"]

[tool.ruff]
target-version = "py312"
extend-exclude = ["libs/tripsphere"]

[tool.ruff.lint]
select = ["E", "F", "B", "I"]

[tool.mypy]
strict = true
python_version = "3.12"

[tool.pyright]
typeCheckingMode = "strict"
include = ["src", "tests"]
